search code added and readme updated
