+++ challengeReadme/depthFirst.md
@@ -1,0 +1,21 @@
+# Depth First Search Graph
+## Code Challenge 37
+
+### Challenge Description
+    - Create a function that accepts an adjacency list as a graph, and conducts a depth first search traversal.
+      without utilizing any of the built-in methods available to your language. Return a collection of nodes in their
+      pre-order depth-first traversal order.
+### Approach & Efficiency
+    
+    Inspiration and help online:
+        - https://www.baeldung.com/java-graphs
+        - https://www.geeksforgeeks.org/graph-and-its-representations/
+        - https://www.geeksforgeeks.org/find-if-there-is-a-path-between-two-vertices-in-a-given-graph/
+        
+### Visual 
+![Route Map](../assets/depthFirstGraph.png)
+
+### Solution
+[Code Solution](/src/main/java/Graph/DepthFirst.java)
+
+[Code Tests](/src/test/java/GraphTest.java)
